render_requirements.txt
discord.py>=2.5.2
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
pillow>=11.3.0
aiosqlite>=0.21.0
email-validator>=2.2.0
psycopg2-binary>=2.9.10